LayoutTests/js/dfg-array-length-dead-expected.txt

Tests that an array length access being dead does not result in register allocation failures.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo({f:[]}) is 42 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
